Fixed pr notes
